requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
groq
scikit-learn
numpy
jinja2
python-multipart
aiosqlite
